refactor(docx-core): split inPlaceModifier into focused modules
Behavior-preserving split of the 3,185-LOC in-place track-change
reconstruction engine into seven focused modules plus an orchestrator
that re-exports the full public surface:

- inPlaceModifier-shared      (revision-id state, DOM utils)
- inPlaceModifier-containers  (container topology + ContainerResolutionError)
- inPlaceModifier-bookmarks   (bookmark-marker survival/cloning)
- inPlaceModifier-wrappers    (w:ins/w:del/move/format wrapping primitives)
- inPlaceModifier-deletion    (deleted/move/field content insertion)
- inPlaceModifier-presplit    (mixed/interleaved run pre-splitting)
- inPlaceModifier-postprocess (grouping, no-op suppression, coalescing)
- inPlaceModifier             (modifyRevisedDocument orchestrator + barrel)

No behavior change: code moved verbatim, public export surface from
./inPlaceModifier.js is byte-identical (27 names, verified against the
pre-split .d.ts). Import graph is an acyclic DAG; the core barrel is the
only importer of the leaf modules.

Lean traceability: replaced the stale inPlaceModifier.ts line-number
citations (already drifted onto blank/comment lines) in
verification/lean/LeanSpike/Spec.lean and lean-spec-bridge.test.ts with
greppable @lean-segment anchors (container-topology, field-wrapper-emission)
so the references survive future splits. lake build green (6586 jobs);
Lean differential + bridge tests pass.

Verified: docx-core build + lint clean, 1302 tests pass, no import cycle.

Closes #318.
